# server/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DeadStockSolution API",
    "version": "1.0.0",
    "description": "Baseline OpenAPI contract for public server routes"
  },
  "paths": {
    "/api/account": {
      "get": {
        "operationId": "getAccount",
        "summary": "Get account details",
        "tags": [
          "Account"
        ],
        "responses": {
          "200": {
            "description": "Account details response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateAccount",
        "summary": "Update account details",
        "tags": [
          "Account"
        ],
        "responses": {
          "200": {
            "description": "Account update response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid account update payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "409": {
            "description": "Account update conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteAccount",
        "summary": "Deactivate account",
        "tags": [
          "Account"
        ],
        "responses": {
          "200": {
            "description": "Account deactivation response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid account deactivation payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/drug-master": {
      "get": {
        "operationId": "getAdminDrugMasterList",
        "summary": "Get admin drug master list",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Admin drug master list response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "500": {
            "description": "Admin drug master list failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/error-codes": {
      "get": {
        "operationId": "getAdminErrorCodes",
        "summary": "Get admin error codes",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Admin error codes response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createAdminErrorCode",
        "summary": "Create admin error code",
        "tags": [
          "Admin"
        ],
        "responses": {
          "201": {
            "description": "Admin error code created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid admin error code payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "500": {
            "description": "Admin error code create failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/log-center": {
      "get": {
        "operationId": "getAdminLogCenter",
        "summary": "Get admin log center entries",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Admin log center response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid admin log center query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/log-center/openclaw": {
      "post": {
        "operationId": "postAdminLogCenterOpenClaw",
        "summary": "Escalate admin log center entry to OpenClaw",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Admin log center OpenClaw escalation response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid admin log center OpenClaw escalation payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "404": {
            "description": "Admin log center entry not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/business-hours": {
      "get": {
        "operationId": "getBusinessHours",
        "summary": "Get my business hours",
        "tags": [
          "BusinessHours"
        ],
        "responses": {
          "200": {
            "description": "Business hours response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "500": {
            "description": "Business hours fetch failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateBusinessHours",
        "summary": "Update my business hours",
        "tags": [
          "BusinessHours"
        ],
        "responses": {
          "200": {
            "description": "Business hours updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid business hours payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "409": {
            "description": "Business hours version conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "500": {
            "description": "Business hours update failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/openclaw/commands": {
      "post": {
        "operationId": "receiveOpenClawCommand",
        "summary": "Receive OpenClaw command webhook",
        "tags": [
          "OpenClaw"
        ],
        "responses": {
          "200": {
            "description": "OpenClaw command processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid OpenClaw command payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "401": {
            "description": "OpenClaw command authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "403": {
            "description": "OpenClaw command rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "500": {
            "description": "OpenClaw command processing failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "503": {
            "description": "OpenClaw command service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/requests": {
      "post": {
        "operationId": "createUserRequest",
        "summary": "Submit user request",
        "tags": [
          "Requests"
        ],
        "responses": {
          "201": {
            "description": "User request created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized request submission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/timeline": {
      "get": {
        "operationId": "getTimeline",
        "summary": "Get timeline events",
        "tags": [
          "Timeline"
        ],
        "responses": {
          "200": {
            "description": "Timeline response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid timeline cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "GenericResponse": {
        "type": "object",
        "additionalProperties": true
      },
...
    }
  }
}